fix compile error

+++ src/core/src/sge_core/graph/Color.h
@@ -62,43 +62,43 @@ constexpr ColorType ColorType_make(ColorModel model, ColorCompressType compress)
 
 #define ColorType_ENUM_LIST(E) \
 	E(None,) \
-	E(Rb,		= ColorType_make(ColorModel::R,		ColorElementType::UNorm8 )) \
-	E(Rs,		= ColorType_make(ColorModel::R,		ColorElementType::UNorm16)) \
-	E(Rh,		= ColorType_make(ColorModel::R,		ColorElementType::Float16)) \
-	E(Rf,		= ColorType_make(ColorModel::R,		ColorElementType::Float32)) \
+	E(Rb,		= (u16)ColorType_make(ColorModel::R,	ColorElementType::UNorm8 )) \
+	E(Rs,		= (u16)ColorType_make(ColorModel::R,	ColorElementType::UNorm16)) \
+	E(Rh,		= (u16)ColorType_make(ColorModel::R,	ColorElementType::Float16)) \
+	E(Rf,		= (u16)ColorType_make(ColorModel::R,	ColorElementType::Float32)) \
 	\
-	E(Lb,		= ColorType_make(ColorModel::L,		ColorElementType::UNorm8 )) \
-	E(Ls,		= ColorType_make(ColorModel::L,		ColorElementType::UNorm16)) \
-	E(Lh,		= ColorType_make(ColorModel::L,		ColorElementType::Float16)) \
-	E(Lf,		= ColorType_make(ColorModel::L,		ColorElementType::Float32)) \
+	E(Lb,		= (u16)ColorType_make(ColorModel::L,	ColorElementType::UNorm8 )) \
+	E(Ls,		= (u16)ColorType_make(ColorModel::L,	ColorElementType::UNorm16)) \
+	E(Lh,		= (u16)ColorType_make(ColorModel::L,	ColorElementType::Float16)) \
+	E(Lf,		= (u16)ColorType_make(ColorModel::L,	ColorElementType::Float32)) \
 	\
-	E(LAb,		= ColorType_make(ColorModel::LA,	ColorElementType::UNorm8 )) \
-	E(LAs,		= ColorType_make(ColorModel::LA,	ColorElementType::UNorm16)) \
-	E(LAh,		= ColorType_make(ColorModel::LA,	ColorElementType::Float16)) \
-	E(LAf,		= ColorType_make(ColorModel::LA,	ColorElementType::Float32)) \
+	E(LAb,		= (u16)ColorType_make(ColorModel::LA,	ColorElementType::UNorm8 )) \
+	E(LAs,		= (u16)ColorType_make(ColorModel::LA,	ColorElementType::UNorm16)) \
+	E(LAh,		= (u16)ColorType_make(ColorModel::LA,	ColorElementType::Float16)) \
+	E(LAf,		= (u16)ColorType_make(ColorModel::LA,	ColorElementType::Float32)) \
 	\
-	E(RGb,		= ColorType_make(ColorModel::RG,	ColorElementType::UNorm8 )) \
-	E(RGs,		= ColorType_make(ColorModel::RG,	ColorElementType::UNorm16)) \
-	E(RGh,		= ColorType_make(ColorModel::RG,	ColorElementType::Float16)) \
-	E(RGf,		= ColorType_make(ColorModel::RG,	ColorElementType::Float32)) \
+	E(RGb,		= (u16)ColorType_make(ColorModel::RG,	ColorElementType::UNorm8 )) \
+	E(RGs,		= (u16)ColorType_make(ColorModel::RG,	ColorElementType::UNorm16)) \
+	E(RGh,		= (u16)ColorType_make(ColorModel::RG,	ColorElementType::Float16)) \
+	E(RGf,		= (u16)ColorType_make(ColorModel::RG,	ColorElementType::Float32)) \
 	\
-	E(RGBb,		= ColorType_make(ColorModel::RGB,	ColorElementType::UNorm8 )) \
-	E(RGBs,		= ColorType_make(ColorModel::RGB,	ColorElementType::UNorm16)) \
-	E(RGBh,		= ColorType_make(ColorModel::RGB,	ColorElementType::Float16)) \
-	E(RGBf,		= ColorType_make(ColorModel::RGB,	ColorElementType::Float32)) \
+	E(RGBb,		= (u16)ColorType_make(ColorModel::RGB,	ColorElementType::UNorm8 )) \
+	E(RGBs,		= (u16)ColorType_make(ColorModel::RGB,	ColorElementType::UNorm16)) \
+	E(RGBh,		= (u16)ColorType_make(ColorModel::RGB,	ColorElementType::Float16)) \
+	E(RGBf,		= (u16)ColorType_make(ColorModel::RGB,	ColorElementType::Float32)) \
 	\
-	E(RGBAb,	= ColorType_make(ColorModel::RGBA,	ColorElementType::UNorm8 )) \
-	E(RGBAs,	= ColorType_make(ColorModel::RGBA,	ColorElementType::UNorm16)) \
-	E(RGBAh,	= ColorType_make(ColorModel::RGBA,	ColorElementType::Float16)) \
-	E(RGBAf,	= ColorType_make(ColorModel::RGBA,	ColorElementType::Float32)) \
+	E(RGBAb,	= (u16)ColorType_make(ColorModel::RGBA,	ColorElementType::UNorm8 )) \
+	E(RGBAs,	= (u16)ColorType_make(ColorModel::RGBA,	ColorElementType::UNorm16)) \
+	E(RGBAh,	= (u16)ColorType_make(ColorModel::RGBA,	ColorElementType::Float16)) \
+	E(RGBAf,	= (u16)ColorType_make(ColorModel::RGBA,	ColorElementType::Float32)) \
 	\
-	E(BC1,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC1 )) \
-	E(BC2,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC2 )) \
-	E(BC3,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC3 )) \
-	E(BC4,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC4 )) \
-	E(BC5,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC5 )) \
-	E(BC6h,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC6h)) \
-	E(BC7,		= ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC7 )) \
+	E(BC1,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC1 )) \
+	E(BC2,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC2 )) \
+	E(BC3,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC3 )) \
+	E(BC4,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC4 )) \
+	E(BC5,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC5 )) \
+	E(BC6h,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC6h)) \
+	E(BC7,		= (u16)ColorType_make(ColorModel::BlockCompression, ColorCompressType::BC7 )) \
 //----
 	SGE_ENUM_CLASS(ColorType, : u16)
